Update unprivileged user information
Add an argument for the unprivileged user's name. Add explicit
variables for the group name and gid that are based on the values for
the user name and uid. Base the home directory on the value provided
for the user name.

Combined these changes allow reasonable build-time modification,
clearly distinguish what values are being used in command calls, and
set reasonable relationships between arguments and environment
variables.