Include the image name as part of the tag
The `--tag` option for `docker buildx imagetools create` expects a full
image name and tag in the form user/app:tag. The example in the Docker
documentation only produces the tag portion which results in the
command attempting to push to the Docker library with the tag as the
image target. Since our token has no such authorization it errors with
ERROR: server message: insufficient_scope: authorization failed